Fix pricing issues (#17)
* Fix pricing updates

* Update lbPair tokenX/Y pricing

* Clean up log
