+++ src/entities/bin.ts
@@ -1,24 +1,17 @@
-import { BigInt, BigDecimal } from "@graphprotocol/graph-ts";
+import { BigInt, BigDecimal, Address } from "@graphprotocol/graph-ts";
 import { Bin, LBPair, Token } from "../../generated/schema";
 import { BIG_DECIMAL_ONE, BIG_DECIMAL_ZERO, BIG_INT_ZERO } from "../constants";
+import { loadToken } from "../entities";
 import { getPriceYOfBin } from "../utils";
 
-export function trackBin(
-  lbPair: LBPair,
-  binId: BigInt,
-  tokenX: Token,
-  tokenY: Token,
-  amountXIn: BigDecimal,
-  amountXOut: BigDecimal,
-  amountYIn: BigDecimal,
-  amountYOut: BigDecimal,
-  minted: BigInt,
-  burned: BigInt
-): Bin {
+export function loadBin(lbPair: LBPair, binId: BigInt): Bin {
   const id = lbPair.id.concat("#").concat(binId.toString());
   let bin = Bin.load(id);
 
   if (!bin) {
+    const tokenX = loadToken(Address.fromString(lbPair.tokenX));
+    const tokenY = loadToken(Address.fromString(lbPair.tokenY));
+
     bin = new Bin(id);
     bin.lbPair = lbPair.id;
     bin.binId = binId;
@@ -28,6 +21,21 @@ export function trackBin(
     bin.priceY = getPriceYOfBin(binId, lbPair.binStep, tokenX, tokenY); // each bin has a determined price
     bin.priceX = BIG_DECIMAL_ONE.div(bin.priceY);
   }
+
+  return bin;
+}
+
+export function trackBin(
+  lbPair: LBPair,
+  binId: BigInt,
+  amountXIn: BigDecimal,
+  amountXOut: BigDecimal,
+  amountYIn: BigDecimal,
+  amountYOut: BigDecimal,
+  minted: BigInt,
+  burned: BigInt
+): Bin {
+  const bin = loadBin(lbPair, binId);
 
   bin.totalSupply = bin.totalSupply.plus(minted).minus(burned);
   bin.reserveX = bin.reserveX.plus(amountXIn).minus(amountXOut);
